new branch work with subdirs flags
